pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "nebula"
version = "0.1.0"
description = "Nebula DNA-only precision wellness backend MVP"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "click>=8.1",
    "python-dateutil>=2.8",
    # Surveillance system uses only stdlib for HTTP (urllib) —
    # no extra network libraries needed for MVP
]

[project.scripts]
nebula = "nebula.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.2",
    "mypy>=1.7",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
